moved projects scripts folder
