+++ Assets/_Application/Projects/AssetUriFactory.cs.meta
