+++ Assets/_Application/Projects/ExtensionMethods/UriExtensions.cs
